Refactored the canary flag feature
- Now the clients are downloaded and updated when needed

+++ build/meterian.sh
@@ -20,6 +20,20 @@ exitWithErrorMessageWhenApiTokenIsUnset() {
 	fi
 }
 
+updateClient() {
+	METERIAN_JAR_PATH=$1
+	CLIENT_TARGET_URL=$2
+
+	LOCAL_CLIENT_LAST_MODIFIED_DATE=$(date -d "$(ls --full-time ${METERIAN_JAR_PATH} | cut -d" " -f6-8)" +%F)
+	REMOTE_CLIENT_LAST_MODIFIED_DATE=$(date -d "$(curl -s -L -I "${CLIENT_TARGET_URL}" \
+									| grep Last-Modified: | cut -d" " -f2-)" +%F)
+	if [[ "${REMOTE_CLIENT_LAST_MODIFIED_DATE}" > "${LOCAL_CLIENT_LAST_MODIFIED_DATE}" ]];
+	then
+		echo Updating the client$(test -n "${CLIENT_CANARY_FLAG}" && echo " canary" || true)...
+		curl -s -o ${METERIAN_JAR_PATH} "${CLIENT_TARGET_URL}"  >/dev/null
+	fi
+}
+
 INDEPENDENT_METERIAN_CLI_OPTIONS="(--version|--help)"
 
 # dump docker packaged version unless '--version' requested
@@ -35,8 +49,16 @@ METERIAN_JAR=/tmp/meterian-cli.jar
 if [[ -n ${CLIENT_CANARY_FLAG}  ]];
 then
 	METERIAN_JAR=/tmp/meterian-cli-canary.jar
-	echo Downloading client canary...
-	curl -s -o ${METERIAN_JAR} "https://www.meterian.io/downloads/meterian-cli-canary.jar"  >/dev/null
+	# download cli-canary if it does not exist 
+	if [[ ! -f ${METERIAN_JAR} ]];
+	then
+		echo Downloading client canary...	
+		curl -s -o ${METERIAN_JAR} "https://www.meterian.io/downloads/meterian-cli-canary.jar"  >/dev/null
+	else
+		# update cli-canary if necessary
+		updateClient "${METERIAN_JAR}" "https://www.meterian.io/downloads/meterian-cli-canary.jar"
+	fi
+	
 else
 	# update the client if necessary
 	if [[ ! -f ${METERIAN_JAR} ]];
@@ -44,14 +66,7 @@ else
 		mv /meterian-cli.jar /tmp/
 	fi
 
-	LOCAL_CLIENT_LAST_MODIFIED_DATE=$(date -d "$(ls --full-time ${METERIAN_JAR} | cut -d" " -f6-8)" +%F)
-	REMOTE_CLIENT_LAST_MODIFIED_DATE=$(date -d "$(curl -s -L -I "https://www.meterian.com/downloads/meterian-cli.jar" \
-									| grep Last-Modified: | cut -d" " -f2-)" +%F)
-	if [[ "${REMOTE_CLIENT_LAST_MODIFIED_DATE}" > "${LOCAL_CLIENT_LAST_MODIFIED_DATE}" ]];
-	then
-		echo Updating the client...
-		curl -s -o ${METERIAN_JAR} "https://www.meterian.com/downloads/meterian-cli.jar"  >/dev/null
-	fi
+	updateClient "${METERIAN_JAR}" "https://www.meterian.com/downloads/meterian-cli.jar"
 fi
 
 # launching the client - note the different launch if version requested to preserve the "--version" base functionality
